Manual merging into structure of #12
